Move paywall custom variable tests
